Convert to embedded-hal 1.0.0-rc.1

+++ src/correction.rs
@@ -10,10 +10,10 @@ use micromath::F32Ext;
 pub fn calculate_raw_threshold_value(it: IntegrationTime, gain: Gain, lux: f32) -> u16 {
     let factor = get_lux_raw_conversion_factor(it, gain);
     if (gain == Gain::OneQuarter || gain == Gain::OneEighth) && lux > 1000.0 {
-        let lux = inverse_high_lux_correction(f32::from(lux));
-        (lux / f32::from(factor)) as u16
+        let lux = inverse_high_lux_correction(lux);
+        (lux / factor) as u16
     } else {
-        (f32::from(lux) / f32::from(factor)) as u16
+        (lux / factor) as u16
     }
 }
